+++ extension/src/test/runTest.ts
@@ -1,5 +1,5 @@
 import * as path from 'path';
-
+// import * as process from 'process';
 import { runTests } from 'vscode-test';
 
 async function main() {
@@ -7,11 +7,14 @@ async function main() {
 		// The folder containing the Extension Manifest package.json
 		// Passed to `--extensionDevelopmentPath`
 		const extensionDevelopmentPath = path.resolve(__dirname, '../../../../');
-
-		// The path to the extension test script
+		let extensionTestsPath ='';
+		// The path to the extension test or code coverage script
 		// Passed to --extensionTestsPath
-		const extensionTestsPath = path.resolve(__dirname, './suite/index');
-
+		if (process.argv[2] === '--codecoverage') {
+		 	extensionTestsPath = path.resolve(__dirname, './suite/codeCoverage');
+		} else {
+			extensionTestsPath = path.resolve(__dirname, './suite/index');
+		}
 		// Download VS Code, unzip it and run the integration test
 		await runTests({ extensionDevelopmentPath, extensionTestsPath });
 	} catch (err) {
