+++ package.json
@@ -615,7 +615,8 @@
 	"scripts": {
 		"compile": "tsc -p ./",
 		"watch": "tsc -p -watch ./",
-		"test": "npm run compile && node ./out/test/runTest.js"
+		"test": "npm run compile && node ./out/test/runTest.js",
+		"cc": "npm run compile && node ./out/test/runTest.js --codecoverage"
 	},
 	"devDependencies": {
 		"@istanbuljs/nyc-config-typescript": "^1.0.1",
